Add files via upload
This file consists of some pages that have been created but still being worked upon. The previous file had initial html and css designs for the pages accessible to the customers. We added the payment and order display on the same page in this new file. Using a css framework, AdminLte.